perf(retrieval): separate semantic and lexical queries
